Show left and right hand values when equality fails
For assertions that contain function calls, displaying just the
expression is not helpful. For e.g.

  Assertion failed
  String.upcase("foo") == "Foo"

The learner is given no extra information to help them learn. Sure they
might consult the documentation (and should!), but perhaps a better
experience would be for us to inform the learner about the values it
encountered.

  Assertion failed
  String.upcase("foo") == "Foo"
  left:  "FOO"
  right: "Foo"

Learner: :bulb: ah hah! it upcases the whole string! :dizzy: